sikio_la_chama_backend/requirements.txt
asgiref==3.9.2
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-extensions==3.2.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
tzdata==2025.2
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
gunicorn==23.0.0
whitenoise==6.7.0
